refactor(app): extract editor input handling to prompt-input/editor-input.ts
Move composing/IME signal, handleBlur, composition handlers, reconcile,
the prompt -> DOM reconcile effect, the carry-over effect, handleInput,
and addPart into createEditorInput factory.

For now atOnInput/slashOnInput/closePopover/resetHistoryNavigation are
passed in as direct closures from PromptInput. The next commit extracts
popover-controllers and converts these to a popovers() ref forward to
break the addPart <-> popover-handlers cycle.

Pure functions (parseEditorToParts, isNormalizedEditor, createPill,
createTextFragment, setRangeEdge) are imported directly from editor-
serialize/editor-dom rather than threaded through deps.
